Allow twoFactorAuthPin with leading zeros
Currently, the `twoFactorAuthPin` is being stored as an integer. When a PIN is entered which has a leading zero, because the PIN is stored as an integer, the leading zero is dropped, but this is done without the user knowing. As a result, when the user tries to use the PIN they believe they entered, Google Home does not accept the PIN.

In my proposed change, I've changed `twoFactorAuthPin`'s `type` to `string`, which I've manually done in the JSON config to confirm it works well with Google Home. 
I have also added a `pattern` which contains regex (which I believe) should ensure users can enter only strings of any length comprised only of numbers from 0-9.